% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CAplot.R
\name{CAplot}
\alias{CAplot}
\title{CAplot run a CA
(with the \code{ExPosition} package)
and generates the standard graphs and tables.
Note: \emph{Still Under Development}.}
\usage{
CAplot(
  resCA,
  data,
  DESIGN = NULL,
  graphs = 12,
  printGraphs = FALSE,
  col4I = NULL,
  col4J = NULL,
  show.TI = FALSE,
  show.CI = TRUE,
  mean.cex = 3,
  mean.textcex = 3,
  display.labels.ind = FALSE,
  display.labels.var = TRUE,
  display.points.mean = TRUE,
  mean.constraints = NULL,
  scale.mean.constraints = 1.5,
  max.n4bar = 40,
  save2pptx = FALSE,
  title4pptx = "CA Results"
)
}
\arguments{
\item{resCA}{Output from epCA}

\item{data}{A data frame or a matrix with
data suitable for a CA.}

\item{DESIGN}{Default: NULL.
A design vector (could be factor or character)
or (Boolean) matrix used to assign rows
to groups.}

\item{graphs}{do we want graphs? At the moment, this parameter is ignored.
Current Default is \code{12} which indicates that
the graphs are generated for the first
2 components. Note that current version
is creating output only for the first two
components.}

\item{printGraphs}{(Default: FALSE)
do we want to print the graphics as \code{.png}?}

\item{col4I}{a color vector for
plotting the rows (if \code{NULL}
Default) use colors from \code{ExPosition::epCA}.}

\item{col4J}{a color vector for
plotting the columns (if \code{NULL}
Default) use colors from \code{ExPosition::epCA}.}

\item{show.TI}{whether to plot the tolerance intervals or not. Default: FALSE}

\item{show.CI}{whether to plot the confidence intervals or not. Default: TRUE}

\item{mean.cex}{the size of the dots of the means. Default: 3}

\item{mean.textcex}{the size of the texts of the means. Default: 3}

\item{display.labels.ind}{If TRUE, the labels of rows will be printed. Default: FALSE.}

\item{display.labels.var}{If TRUE, the labels of columns will be printed. Default: TRUE.}

\item{display.points.mean}{If TRUE, the mean row factor scores will be plotted. Default: TRUE.}

\item{mean.constraints}{A list of the constraints (that include \code{minx}, \code{miny}, \code{maxx}, and \code{maxy})
The constraints of the figure that only includes the means. This constraints
will be used if \code{only.mean = TRUE}. Default: NULL}

\item{scale.mean.constraints}{A value used to scale the constraints (by multiplication).
This function is used to adjust the constraints when the confidence or the tolerance intervals are outside of the figure.
Default: 1.5}

\item{max.n4bar}{When the number of bars exceed this value, the labels will be hidden. Default: 40.}

\item{save2pptx}{Default: FALSE}

\item{title4pptx}{Title of the PPTX, Default:
'CA Results'.}
}
\value{
A list made of two lists
}
\description{
Generates the standard graphs and tables for CA.
}
\details{
Work in Progress
}
\seealso{
\code{\link[ExPosition]{epCA}}
\code{\link[PTCA4CATA]{PlotScree}}, \code{\link[PTCA4CATA]{createFactorMap}}
}
\author{
Luke Moraglia and Hervé Abdi
}
