% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveResults2xlsx.R
\name{saveCAResults2xlsx}
\alias{saveCAResults2xlsx}
\title{saveCAResults2xlsx}
\usage{
saveCAResults2xlsx(
  results.stats,
  data_ca,
  design = NULL,
  file2Save.xlsx = "CA Results.xlsx"
)
}
\arguments{
\item{results.stats}{The same \code{results.stats} from \code{CAplot()}}

\item{data_ca}{The data passed to \code{epCA()}}

\item{design}{The design variable passed to \code{epCA()}}

\item{file2Save.xlsx}{File name / path to save the Excel sheet}
}
\value{
a list of data written to the Excel sheet, with class \code{"save2xlsx"}.
}
\description{
Save an Excel Spreadsheet based on the \code{results.stats}
from \code{\link[=CAplot]{CAplot()}}
}
\author{
Luke Moraglia
}
