% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveResults2xlsx.R
\name{savePCAResults2xlsx}
\alias{savePCAResults2xlsx}
\title{savePCAResults2xlsx}
\usage{
savePCAResults2xlsx(
  results.stats,
  data_pca,
  design = NULL,
  file2Save.xlsx = "PCA Results.xlsx"
)
}
\arguments{
\item{results.stats}{The same \code{results.stats} from \code{PCAplot()}}

\item{data_pca}{The data passed to \code{epPCA()}}

\item{design}{The design variable passed to \code{epPCA()}}

\item{file2Save.xlsx}{File name / path to save the Excel sheet}
}
\value{
a list of data written to the Excel sheet, with class \code{"save2xlsx"}.
}
\description{
Save an Excel Spreadsheet based on the \code{results.stats}
from \code{\link[=PCAplot]{PCAplot()}}
}
\author{
Luke Moraglia
}
