% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveResults2xlsx.R
\name{saveResultsList2xlsx}
\alias{saveResultsList2xlsx}
\title{saveResultsList2xlsx}
\usage{
saveResultsList2xlsx(list2Save, file2Save.xlsx = "Results.xlsx")
}
\arguments{
\item{list2Save}{a named list of data frames to save}

\item{file2Save.xlsx}{file name to save the .xlsx file}
}
\value{
a list of data written to the Excel sheet, with class \code{"save2xlsx"}.
}
\description{
save a named list of data frames to sheets in an .xlsx file
}
\author{
Luke Moraglia
}
